Show Tasks list of User when logged in. Move timer logic to Timer component. Remove BreakSetter and SessionSetter. Later, they will be set in Settings page for each task.
