+++ client/src/App.js
@@ -1,104 +1,25 @@
-import React, {useRef, useState} from 'react';
+import React, {useState, useEffect} from 'react';
 import {ThemeProvider} from 'styled-components';
 import axios from 'axios';
 import {BrowserRouter as Router, Route, Link, Redirect} from 'react-router-dom';
 
-import useInterval from './customHooks/useInterval';
-
-import BreakSetter from './components/BreakSetter';
-import SessionSetter from './components/SessionSetter';
 import Timer from './components/Timer';
 
 import {Row} from './components/styled-elements';
 import {theme} from './theme';
+import config from './config';
 
-function App({currentUser}) {
-  const [breakLength, setBreakLength] = useState(_minuteToSeconds(5));
-  const [sessionLength, setSessionLength] = useState(_minuteToSeconds(25));
-  const [timeLeft, setTimeLeft] = useState(_minuteToSeconds(25));
-  const [isRunning, setIsRunning] = useState(false);
-  const [isBreak, setIsBreak] = useState(false);
-  const beepRef = useRef();
+function App(props) {
+  const {isAuthenticated, currentUser} = props;
+  const [tasks, setTasks] = useState([]);
 
-  function incBreakLength() {
-    _withinLengthLimit(setBreakLength, breakLength, _minuteToSeconds(1));
-  }
-
-  function decBreakLength() {
-    _withinLengthLimit(setBreakLength, breakLength, _minuteToSeconds(-1));
-  }
-
-  function incSessionLength() {
-    if (!isRunning) {
-      _withinLengthLimit(setTimeLeft, timeLeft, _minuteToSeconds(1));
+  useEffect(() => {
+    if (isAuthenticated) {
+      axios.get(`${config.API_ROOT}/tasks`, {withCredentials: true})
+        .then(response => setTasks(response.data));
     }
-    _withinLengthLimit(setSessionLength, sessionLength, _minuteToSeconds(1));
-  }
+  }, [isAuthenticated]);
 
-  function decSessionLength() {
-    if (!isRunning) {
-      _withinLengthLimit(setTimeLeft, timeLeft, _minuteToSeconds(-1));
-    }
-    _withinLengthLimit(setSessionLength, sessionLength, _minuteToSeconds(-1));
-  }
-
-  function toggleIsRunning() {
-    setIsRunning(!isRunning);
-  }
-
-  function reset() {
-    _stopBeep();
-    setTimeLeft(_minuteToSeconds(25));
-    setBreakLength(_minuteToSeconds(5));
-    setSessionLength(_minuteToSeconds(25));
-    setIsRunning(false);
-    setIsBreak(false);
-  }
-
-  useInterval(_tick, isRunning ? 1000 : null);
-
-
-  function _minuteToSeconds(minute) {
-    return minute * 60;
-  }
-
-  function _withinLengthLimit(setFunction, value, change) {
-    const MAX_LENGTH = 3600;
-    const MIN_LENGTH = 60;
-    const newValue = value + change;
-    if (MIN_LENGTH <= newValue && newValue <= MAX_LENGTH) {
-      setFunction(newValue);
-    }
-  }
-
-  function _tick() {
-    const timeLeftNow = timeLeft - 1;
-    if (timeLeftNow < 0) {
-      _changeSessionTo(!isBreak);
-    } else {
-      if (timeLeftNow === 0) _playBeepFromStart();
-      setTimeLeft(timeLeftNow);
-    }
-  }
-
-  function _changeSessionTo(nowIsBreak) {
-    setIsBreak(nowIsBreak);
-    setTimeLeft(nowIsBreak ? breakLength : sessionLength);
-  }
-
-  function _getTimerLabel() {
-    return isBreak ? 'Break' : 'Work';
-  }
-
-  function _playBeepFromStart() {
-    _stopBeep();
-    beepRef.current.play();
-  }
-
-  function _stopBeep() {
-    beepRef.current.pause();
-    beepRef.current.currentTime = 0;
-  }
 
   return (
     <ThemeProvider theme={theme}>
@@ -111,14 +32,8 @@ function App({currentUser}) {
           }
         </Row>
         <Row>
-          <Timer timeLeft={timeLeft} reset={reset} startStop={toggleIsRunning} timerLabel={_getTimerLabel()}/>
+          <Timer tasks={tasks}/>
         </Row>
-        <Row>
-          <BreakSetter length={breakLength} inc={incBreakLength} dec={decBreakLength}/>
-          <SessionSetter length={sessionLength} inc={incSessionLength} dec={decSessionLength}/>
-        </Row>
-        <audio id="beep" ref={beepRef}
-               src="https://docs.google.com/uc?export=download&id=177Le-I9Z4arIsILN9xicG7-GkGt09PdM"/>
       </div>
     </ThemeProvider>
   );
@@ -131,7 +46,7 @@ function AppWrapper({initialUser}) {
   function authenticate(username, password) {
     console.log('in authenticate');
     axios.post(
-      'http://localhost:3003/api/pomodoro/login',
+      `${config.API_ROOT}/login`,
       {
         username,
         password
@@ -158,9 +73,12 @@ function AppWrapper({initialUser}) {
 
   return (
     <Router>
-      <Route exact path="/" render={props => <App {...props} currentUser={currentUser}/>}/>
-      <Route path="/login" render={props => <LoginComponent {...props} isAuthenticated={isAuthenticated}
-                                                            authenticate={authenticate}/>}/>
+      <Route exact path="/" render={
+        props => <App {...props} isAuthenticated={isAuthenticated} currentUser={currentUser}/>
+      }/>
+      <Route path="/login" render={
+        props => <LoginComponent {...props} isAuthenticated={isAuthenticated} authenticate={authenticate}/>
+      }/>
     </Router>
   );
 }
