+++ client/src/config.js
@@ -1,0 +1,1 @@
+exports.API_ROOT = 'http://localhost:3003/api/pomodoro';
